requirements.txt
celery==5.2.7
Django==4.0.5
django-celery-beat==2.3.0
django-filter==22.1
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
Faker==13.13.0
pendulum==2.1.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
redis==4.3.4
requests==2.28.0
